requirements.txt
alembic==1.7.5
blinker==1.4
cffi==1.15.0
click==8.0.3
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
pycparser==2.21
python-dotenv==0.19.2
SQLAlchemy==1.4.29
Werkzeug==2.0.2
WTForms==3.0.1
